Add support for DM6 OO and ForceDate
For the rest of the changes coming, we will need to use the ForceDate option of
Date::Manip to make sure the date context is setup correctly using the
now_callback. Unfortunately, with the functional interface, you can only set
ForceDate globally (not per-call). It would be better, therefore, to use the OO
interface which allows per-object configuration. However, this is only available
in 6.00+, so add a helper method to bridge either interface.

When using the older functional interface (either because the developer uses a
version of Date::Manip older than 6.00 or because they are using 6.14+ with
backend set to "DM5"), we will record the previous ForceDate setting, force the
date while we make the call, and then flip it back to the previous setting.

If the version of Date::Manip can be tested with DM5 turned one, this adds a
test to make sure the ForceDate setting is reverted after each use.