Add support for explicit date ranges
This is implemented in terms of parsing the range recursively, so the
performance of this is not as good as it probably could be since we end up
constructing additional dates that we throw away. However, it does the job and,
unless I'm mistaken, the parsing should proceed in such a way that you will
always reach a base case with any input.